examples/03_lectura_tarot/requirements.txt
lunita-mini
python-dotenv
